requirements.txt
defusedxml==0.5.0
Django==2.1.4
django-filter==2.0.0
django-taggit==0.23.0
djangorestframework==3.9.0
djangorestframework-xml==1.4.0
gunicorn==19.6.0
Markdown==3.0.1
#mod-wsgi==4.6.5
mysqlclient==1.3.14
pytz==2018.7
